polyglossia + pygmentex + fonts
Use polyglossia to setup the default language to Spanish but with the
possibility to write text in English and Russian.

Set FreeSerif, FreeSans and FreeMono as the default fonts; they support
mathematical symbols and Cyrillic characters.

Set pygmentex to use it in replace of lstlisting. It requires support
from outside (call pygmentex.py)