Call pygmenter.py after first xelatex round
latexmk call several time to the Tex engine to generate the final
document.

On the first round, pygmentex will generate a .snippets file that will
have all the code snippets to be highlighted.

This highlighting process is 'out of band': it must be done by the
scripts/pygmenter.py program after the first Tex engine round but before
any other further pass.

For that we created our own Tex engine: scripts/pygmenter_xelatex which
calls xelatex but it also calls scripts/pygmenter.py

The proper hook is set on latexmk to make it work.