Impl asciidiagram flag for src code
The asciidiagram flag will make the highlighted source code unbreakable.
That's it, it will not be break by a new page which it is important when
doing ascii diagrams.

Also, asciidiagram changes the style of the highlighted code to
candombeascii. This style behaves the same as the origin candombe style
but it makes the spaces in the code to have a the same width that any
other character.

Change the candombe style to not draw any decoration for the whitespaces
and syntax errors.